requirements.txt

sentence-transformers
langchain
transformers
faiss-cpu
matplotlib
pandas
python-dotenv
streamlit
PyPDF2
python-docx